+++ src/entities/profile/api/editprofile.ts
@@ -1,5 +1,4 @@
 import { api } from "@shared/api";
-import { ApiResponse } from "@shared/types";
 import { MemberProfile } from "@/entities/profile/api";
 
 export interface UpdateMemberRequest {
@@ -7,6 +6,12 @@ export interface UpdateMemberRequest {
   tagIds: number[];
 }
 
+export interface EditProfileParams {
+  nickname: string;
+  tagIds: number[];
+}
+
 export const editProfileApi = {
-  updateMemberProfile: (body: UpdateMemberRequest) => api.patch<MemberProfile>("/member/me", body),
+  updateMemberProfile: (body: UpdateMemberRequest) =>
+    api.patch<MemberProfile>("/member/me", body),
 };
