+++ src/entities/profile/api/index.ts
@@ -1,4 +1,4 @@
 export { memberApi } from "./userinfo";
 export type { MemberProfile } from "./userinfo";
 export { editProfileApi } from "./editprofile";
-export type { UpdateMemberRequest } from "./editprofile";
+export type { UpdateMemberRequest, EditProfileParams } from "./editprofile";
